Do not expect STDERR in all Runners
This one was silly. The test was in command_line_spec, which is to
test thing common to all Runners. But as I pointed out in the
previous commit, only Posix and Process are capable of returning
things from STDERR. So, of course, it fails in JRuby which isn't
capable of running either Runner.

This change moves the test to those Runners' specific tests.